area/ui: Fix Can't resolve 'next/router' error message
