+++ ui/package.json
@@ -25,6 +25,7 @@
     "classnames": "^2.3.1",
     "date-fns": "^2.28.0",
     "google-protobuf": "^3.19.0",
+    "next": "^12.0.7",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "tailwindcss": "^3.0.1"
