+++ ui/packages/app/web/src/pages/index.tsx
@@ -1,34 +1,9 @@
 import {QueryServiceClient} from '@parca/client';
 import {useLocation, useNavigate} from 'react-router-dom';
 import ProfileExplorer from '../components/ProfileExplorer';
+import {parseParams, convertToQueryParams} from '@parca/functions';
 
 const apiEndpoint = process.env.REACT_APP_PUBLIC_API_ENDPOINT;
-
-const transformToArray = params => params.split(',');
-
-const parseParams = (querystring: string) => {
-  const params = new URLSearchParams(querystring);
-
-  const obj: any = {};
-  for (const key of params.keys()) {
-    if (params.getAll(key).length > 1) {
-      obj[key] = params.getAll(key);
-    } else {
-      if (params.get(key).includes(',')) {
-        obj[key] = transformToArray(params.get(key));
-      } else {
-        obj[key] = params.get(key);
-      }
-    }
-  }
-
-  return obj;
-};
-
-const convertToQueryParams = params =>
-  Object.keys(params)
-    .map(key => key + '=' + params[key])
-    .join('&');
 
 const Profiles = () => {
   const location = useLocation();
